Allow server to be offline during validation
If a server is offline during startup or a config reload, instead of
failing, we will log the error. Then, when a client tries to connec to a
particular pool, pgcat will attempt to connect to the upstream database
server and obtain the server info params before passing them to the
client. These server info params will then be cached so the next client
will not need to wait for pgcat to connect to the upstream database
again. These parameters are warmed or re-warmed during a reload.